Add edit and clear modules
